refactor : user item
- refactor user item Controller and Service
