+++ src/main/java/com/projectsassy/sassy/item/repository/ItemRepository.java
@@ -1,10 +1,7 @@
 package com.projectsassy.sassy.item.repository;
 
-import com.projectsassy.sassy.item.domain.Badge;
 import com.projectsassy.sassy.item.domain.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
 import java.util.List;
